app/requirements.txt
django
djangorestframework
markdown
django-filter
psycopg2-binary
django-probes
djangorestframework_simplejwt
drf-spectacular
gunicorn==20.0.4